Fix feature labels in docs
