+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "vtubestudio"
-version = "0.2.0"
+version = "0.3.0-alpha.0"
 edition = "2018"
 authors = ["Walfie <walfington@gmail.com"]
 license = "MIT"
