+++ README.md
@@ -12,7 +12,7 @@ A library for interacting with the [VTube Studio API].
 
 This example creates a `Client` using the provided builder, which:
 
-* connects to `ws://localhost:8001` using [tokio_tungstenite](https://docs.rs/tokio_tungstenite)
+* connects to `ws://localhost:8001` using [`tokio_tungstenite`](https://docs.rs/tokio_tungstenite)
 * authenticates with an existing token (if present and valid)
 * reconnects when disconnected, and retries the failed request on reconnection success
 * requests a new auth token on receiving an auth error, and retries the initial failed
