+++ src/lib.rs
@@ -22,7 +22,7 @@
 //!
 //! This example creates a [`Client`] using the provided [builder](ClientBuilder), which:
 //!
-//! * connects to `ws://localhost:8001` using [tokio_tungstenite](https://docs.rs/tokio_tungstenite)
+//! * connects to `ws://localhost:8001` using [`tokio_tungstenite`](https://docs.rs/tokio_tungstenite)
 //! * authenticates with an existing token (if present and valid)
 //! * reconnects when disconnected, and retries the failed request on reconnection success
 //! * requests a new auth token on receiving an auth error, and retries the initial failed
@@ -86,6 +86,17 @@
 //! [`Response`]: crate::data::Response
 //! [`RequestEnvelope`]: crate::data::RequestEnvelope
 //! [`ResponseEnvelope`]: crate::data::ResponseEnvelope
+//!
+//! # Optional features
+//!
+//! By default, the `tokio-tungstenite` feature is enabled, which includes helpers related to the
+//! [`tokio_tungstenite`] websocket library. This can be disabled in your `Cargo.toml` with
+//! `default-features = false`:
+//!
+//! ```toml
+//! [dependencies]
+#![doc = concat!("vtubestudio = { version = \"", env!("CARGO_PKG_VERSION"), "\", default-features = false }")]
+//! ```
 
 /// Utilities for creating [`Client`]s.
 pub mod client;
